Refactor: Rename api/ to api/v1/
